Take Eugen's feedback into account
