+++ src/components/renderers/controls/array/array-directive.ts
@@ -4,7 +4,8 @@ import {IPathResolver} from '../../../services/pathresolver/jsonforms-pathresolv
 import {PathUtil} from '../../../services/pathutil';
 import {IUISchemaGenerator} from '../../../generators/generators';
 import {AbstractControl, ControlRendererTester} from '../abstract-control';
-import {IUISchemaElement, IGroup, SchemaArray} from '../../../../jsonforms';
+import {IUISchemaElement, IGroup} from '../../../../uischema';
+import {SchemaArray} from "../../../../jsonschema";
 
 class ArrayReadOnlyDirective implements ng.IDirective {
     restrict = 'E';
