+++ src/components/renderers/controls/string/string-directive.ts
@@ -1,7 +1,8 @@
 import {RendererTester, NOT_FITTING} from '../../renderer-service';
 import {IPathResolver} from '../../../services/pathresolver/jsonforms-pathresolver';
 import {AbstractControl, ControlRendererTester} from '../abstract-control';
-import {IUISchemaElement} from '../../../../jsonforms';
+import {IUISchemaElement} from "../../../../uischema";
+
 
 class StringDirective implements ng.IDirective {
     restrict = 'E';
