+++ src/components/renderers/layouts/abstract-layout.ts
@@ -1,7 +1,7 @@
 import {Services, ServiceId} from '../../services/services';
 import {IRuleService, IRuleServiceCallBack} from '../../services/rule/rule-service';
 import {IDataProvider} from '../../services/data/data-service';
-import {ILayout, IRule} from '../../../jsonforms';
+import {ILayout, IRule} from "../../../uischema";
 
 export abstract class AbstractLayout implements IRuleServiceCallBack {
