+++ src/components/renderers/layouts/categories/categorization-directive.ts
@@ -1,7 +1,7 @@
 import {RendererTester, NOT_FITTING} from '../../renderer-service';
 import {IPathResolver} from '../../../services/pathresolver/jsonforms-pathresolver';
 import {AbstractLayout} from '../abstract-layout';
-import {IUISchemaElement} from '../../../../jsonforms';
+import {IUISchemaElement} from "../../../../uischema";
 
 class CategorizationDirective implements ng.IDirective {
     restrict = 'E';
