+++ src/components/renderers/layouts/group/group-directive.ts
@@ -1,7 +1,8 @@
 import {RendererTester, NOT_FITTING} from '../../renderer-service';
 import {IPathResolver} from '../../../services/pathresolver/jsonforms-pathresolver';
 import {AbstractLayout} from '../abstract-layout';
-import {IUISchemaElement} from '../../../../jsonforms';
+import {IUISchemaElement} from "../../../../uischema";
+
 class GroupDirective implements ng.IDirective {
     restrict = 'E';
     templateUrl = 'group.html';
